+++ tests/testthat/test-collect.R
@@ -2,11 +2,11 @@ test_that("basic behavior works", {
   pl_iris <- as_polars_df(iris)
   pl_iris_lazy <- as_polars_lf(iris)
 
-  expect_s3_class(collect(pl_iris_lazy), "data.frame")
+  expect_s3_class(collect(pl_iris_lazy), "tbl_df")
 
   expect_equal(
     collect(pl_iris_lazy),
-    iris
+    as_tibble(iris)
   )
 
   # TODO: droplevels() shouldn't be needed
@@ -16,7 +16,8 @@ test_that("basic behavior works", {
       collect() |>
       droplevels(),
     iris[iris$Species == "setosa", ] |>
-      droplevels(),
+      droplevels() |>
+      as_tibble(),
     ignore_attr = FALSE
   )
 })
@@ -50,6 +51,6 @@ test_that("conversion arguments are used", {
   dat <- pl$LazyFrame(x = 1, y = 2)$cast(x = pl$UInt8, y = pl$Int64)
   expect_equal(
     collect(dat, uint8 = "raw", int64 = "character"),
-    data.frame(x = as.raw(1), y = "2")
+    tibble(x = as.raw(1), y = "2")
   )
 })